chore(array-utils): separate plugins by purpose
