+++ array-utils/src/plugins/enumeration/EnvInfoPlugin.ts
@@ -1,5 +1,5 @@
-import { PluginBase } from "./PluginBase";
-import { EnvDetails } from "../types/index";
+import { PluginBase } from "../PluginBase";
+import { EnvDetails } from "../../types/index";
 
 export class EnvInfoPlugin extends PluginBase {
     constructor() {
