+++ array-utils/src/plugins/funny/index.ts
@@ -1,0 +1,10 @@
+import { CatsPlugin } from "./CatsPlugin";
+import { WcPlugin } from "./WcPlugin";
+import { HeyInternetPlugin } from "./HeyInternetPlugin";
+import { PluginBase } from "../PluginBase";
+
+export const funPlugins: Array<PluginBase> = [
+    new CatsPlugin(),
+    new WcPlugin(),
+    new HeyInternetPlugin()
+];
